transmission matrix
